# api/internal/api/spec/openapi.yaml
openapi: 3.1.0
info:
  title: lifeai API
  version: 0.1.0
  description: |
    Food, training, body metrics, recipes and an AI coach — one record of a life.

    Authenticate with a session token from `/api/auth/login`, or a personal API
    token created at `/api/tokens` (prefix `lai_`), sent as `Authorization: Bearer …`.
    Read tokens can call GET endpoints; write tokens can call everything except
    token management and integration settings.
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        code: { type: string }
    Metrics:
      type: object
      properties:
        weight_kg: { type: number, nullable: true }
        body_fat_pct: { type: number, nullable: true }
        resting_hr: { type: integer, nullable: true }
        sleep_hours: { type: number, nullable: true }
        steps: { type: integer, nullable: true }
        water_ml: { type: integer, nullable: true }
        mood: { type: integer, nullable: true, minimum: 1, maximum: 5 }
        energy: { type: integer, nullable: true, minimum: 1, maximum: 5 }
        note: { type: string }
    Meal:
      type: object
      properties:
        id: { type: integer }
        date: { type: string, format: date }
        name: { type: string }
        slot: { type: string, enum: [breakfast, lunch, dinner, snack] }
        kcal: { type: number }
        protein_g: { type: number }
        carbs_g: { type: number }
        fat_g: { type: number }
        source: { type: string }
        estimate_status: { type: string }
    Recipe:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        summary: { type: string }
        minutes: { type: integer }
        servings: { type: integer }
        kcal_per_serving: { type: number }
        protein_g: { type: number }
        carbs_g: { type: number }
        fat_g: { type: number }
        ingredients: { type: array, items: { type: string } }
        steps: { type: array, items: { type: string } }
        tags: { type: array, items: { type: string } }
        favourite: { type: boolean }
    Workout:
      type: object
      properties:
        id: { type: integer }
        date: { type: string, format: date }
        kind: { type: string }
        activity: { type: string }
        minutes: { type: integer }
        kcal: { type: number, nullable: true }
        distance_km: { type: number, nullable: true }
        avg_hr: { type: integer, nullable: true }
        notes: { type: string }
        started_at: { type: string, nullable: true }
        source: { type: string }
        sources: { type: array, items: { type: string } }
        strava_id: { type: string, description: Set once the session exists on Strava }
        exercises: { type: array, items: { $ref: '#/components/schemas/WorkoutExercise' } }
    WorkoutInput:
      type: object
      required: [minutes]
      properties:
        date: { type: string, format: date, description: Default today }
        kind: { type: string, enum: [strength, cardio, walk, run, cycle, swim, yoga, hiit, sport, other], description: Default strength when exercises are given }
        activity: { type: string }
        minutes: { type: integer, minimum: 1, maximum: 1440 }
        started_at: { type: string, format: date-time }
        kcal: { type: number }
        distance_km: { type: number }
        avg_hr: { type: integer }
        notes: { type: string }
        exercises: { type: array, items: { $ref: '#/components/schemas/WorkoutExercise' } }
    WorkoutExercise:
      type: object
      description: One movement. Give a catalog slug (GET /exercises), or a name; a recognised name is matched to the catalog.
      properties:
        slug: { type: string }
        name: { type: string }
        notes: { type: string }
        sets:
          type: array
          items:
            type: object
            properties:
              reps: { type: integer, nullable: true }
              weight: { type: number, nullable: true, description: Per dumbbell, or the single dumbbell for one-dumbbell movements }
              unit: { type: string, enum: [lb, kg], default: lb }
              seconds: { type: integer, nullable: true }
    WeekTraining:
      type: object
      properties:
        from: { type: string, format: date }
        to: { type: string, format: date }
        minutes: { type: integer }
        goal: { type: integer, nullable: true }
        sessions: { type: integer }
        days: { type: array, items: { type: object, properties: { date: { type: string }, minutes: { type: integer } } } }
        planned:
          type: array
          description: The schedule's fixed sessions this week
          items:
            type: object
            properties:
              date: { type: string, format: date }
              title: { type: string }
              kind: { type: string }
              start: { type: string }
              end: { type: string }
              minutes: { type: integer }
              status: { type: string, enum: [done, today, upcoming, missed] }
              workout_id: { type: integer, nullable: true }
        scheduled_left: { type: integer, description: Minutes of fixed sessions still to come }
        freeform_needed: { type: integer, description: Minutes the goal still needs beyond the fixed sessions left }
    WaterEntry:
      type: object
      properties:
        id: { type: integer }
        amount_ml: { type: number }
        created_at: { type: string }
    Day:
      type: object
      properties:
        date: { type: string, format: date }
        is_today: { type: boolean }
        metrics: { $ref: '#/components/schemas/Metrics' }
        water: { type: array, items: { $ref: '#/components/schemas/WaterEntry' } }
        meals: { type: array, items: { $ref: '#/components/schemas/Meal' } }
        workouts: { type: array, items: { $ref: '#/components/schemas/Workout' } }
        week: { $ref: '#/components/schemas/WeekTraining' }
        totals:
          type: object
          properties:
            kcal: { type: number }
            protein_g: { type: number }
            carbs_g: { type: number }
            fat_g: { type: number }
            workout_minutes: { type: integer }
            meditation_minutes: { type: integer }
paths:
...
  /auth/login:
    post:
      security: []
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: Token and user }
        '401': { description: Bad credentials, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /days/{date}/water:
    parameters:
      - { name: date, in: path, required: true, schema: { type: string, format: date } }
    post:
      summary: Add a drink to the daily total
      description: Reuse the request_id when retrying the same drink, including after an undo. Gallons and fluid ounces use US units.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, unit, request_id]
              properties:
                amount: { type: number, exclusiveMinimum: 0 }
                unit: { type: string, enum: [gal, oz, l, ml] }
                request_id: { type: string, minLength: 8, maxLength: 100 }
      responses:
        '200': { description: Updated day, content: { application/json: { schema: { $ref: '#/components/schemas/Day' } } } }
        '400': { description: Invalid amount or unit }
        '409': { description: Request ID already used for a different drink }
  /days/{date}/water/{waterID}:
    parameters:
      - { name: date, in: path, required: true, schema: { type: string, format: date } }
      - { name: waterID, in: path, required: true, schema: { type: integer } }
    delete:
      summary: Undo a drink and recalculate the daily total
      responses:
        '200': { description: Updated day, content: { application/json: { schema: { $ref: '#/components/schemas/Day' } } } }
        '404': { description: Drink not found }
...
  /days:
    get:
      summary: One summary per date in a range
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        '200': { description: OK }
  /days/{date}:
    parameters:
      - { name: date, in: path, required: true, schema: { type: string }, description: YYYY-MM-DD or "today" }
    get:
      summary: One date in full
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Day' } } } }
    patch:
      summary: Record body metrics
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Metrics' }
      responses:
        '200': { description: OK }
...
  /stats:
    get:
      summary: Trends over a window
      parameters:
        - { name: days, in: query, schema: { type: integer, default: 90 } }
      responses:
        '200': { description: OK }
...
  /meals/{mealID}:
    parameters:
      - { name: mealID, in: path, required: true, schema: { type: integer } }
    patch:
      summary: Edit a meal
      responses:
        '200': { description: OK }
    delete:
      summary: Delete a meal
      responses:
        '200': { description: OK }
  /meals/{mealID}/estimate:
    post:
      summary: Re-run a failed photo estimate
      parameters:
        - { name: mealID, in: path, required: true, schema: { type: integer } }
      responses:
        '202': { description: Queued }
  /recipes:
    get:
      summary: The recipe library
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: tag, in: query, schema: { type: string } }
        - { name: favourite, in: query, schema: { type: boolean } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Recipe' } } } } }
    post:
      summary: Add a recipe
      responses:
        '201': { description: Created }
  /recipes/{recipeID}:
    parameters:
      - { name: recipeID, in: path, required: true, schema: { type: integer } }
    get:
      summary: One recipe
      responses:
        '200': { description: OK }
    patch:
      summary: Edit a recipe
      responses:
        '200': { description: OK }
    delete:
      summary: Delete a recipe
      responses:
        '200': { description: OK }
  /recipes/{recipeID}/cook:
    post:
      summary: Log a meal from a recipe
      parameters:
        - { name: recipeID, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                date: { type: string, format: date }
                slot: { type: string }
                servings: { type: number }
      responses:
        '201': { description: Created }
  /workouts:
    get:
      summary: List sessions newest first, with exercises and sets
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: kind, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 100 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Workout' } } } } }
    post:
      summary: Log a workout, with exercises and sets for strength work
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/WorkoutInput' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Workout' } } } }
  /workouts/{workoutID}:
    parameters:
      - { name: workoutID, in: path, required: true, schema: { type: integer } }
    get:
      summary: One session
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Workout' } } } }
    put:
      summary: Replace a session's details and exercises
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/WorkoutInput' } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Workout' } } } }
    delete:
      summary: Delete a workout
      responses:
        '200': { description: OK }
  /workouts/{workoutID}/strava:
    post:
      summary: Post the session to Strava, or update the activity it is linked to
      description: Requires a Strava connection that granted activity uploads (status can_upload). Returns 403 strava_scope when it must be reconnected.
      parameters:
        - { name: workoutID, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: 'OK: {strava_id, url, created, workout}' }
  /exercises:
    get:
      summary: The exercise catalog with muscles, form steps, cues and demo images
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string, enum: [warmup, legs, push, pull, arms, core] } }
        - { name: muscle, in: query, schema: { type: string } }
      responses:
        '200': { description: 'OK: {categories, muscles, exercises}' }
  /exercises/{slug}/history:
    get:
      summary: Recent sessions of one exercise, newest first
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string }, description: 'Use "custom" with ?name= for a movement outside the catalog' }
        - { name: name, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 10 } }
      responses:
        '200': { description: OK }
  /training/week:
    get:
      summary: Exercise minutes for the Monday-to-Sunday week holding a date, against the weekly goal
      parameters:
        - { name: date, in: query, schema: { type: string, format: date } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/WeekTraining' } } } }
  /training/schedule:
    get:
      summary: 'The weekly schedule: fixed sessions and notes on freeform training ({schedule: null} when unset)'
      responses:
        '200': { description: OK }
    put:
      summary: Replace the weekly schedule
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                notes: { type: string }
                fixed:
                  type: array
                  items:
                    type: object
                    required: [title, days, start, end]
                    properties:
                      key: { type: string }
                      title: { type: string }
                      kind: { type: string, default: strength }
                      days: { type: array, items: { type: integer, minimum: 0, maximum: 6 }, description: 0 = Sunday }
                      start: { type: string, example: '06:30' }
                      end: { type: string, example: '07:30' }
      responses:
        '200': { description: OK }
...
  /meditations/{meditationID}:
    delete:
      summary: Delete a meditation
      parameters:
        - { name: meditationID, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: OK }
  /journal:
    get:
      summary: Journal entries, newest first
      parameters:
        - { name: q, in: query, schema: { type: string } }
      responses:
        '200': { description: OK }
    post:
      summary: Write an entry
      responses:
        '201': { description: Created }
  /journal/{entryID}:
    parameters:
      - { name: entryID, in: path, required: true, schema: { type: integer } }
    patch:
      summary: Edit an entry
      responses:
        '200': { description: OK }
    delete:
      summary: Delete an entry
      responses:
        '200': { description: OK }
  /photos:
    get:
      summary: Photos, newest first
      parameters:
        - { name: kind, in: query, schema: { type: string, enum: [progress, food, ingredients] } }
        - { name: pose, in: query, schema: { type: string, enum: [front, side, back] } }
      responses:
        '200': { description: OK }
    post:
      summary: Upload a photo (multipart, field "file")
      responses:
        '201': { description: Created }
  /photos/{photoID}/file:
    get:
      summary: The image bytes
      parameters:
        - { name: photoID, in: path, required: true, schema: { type: integer } }
        - { name: size, in: query, schema: { type: string, enum: [thumb] } }
      responses:
        '200': { description: Image }
...
  /tokens/{tokenID}:
    delete:
      summary: Revoke a token
      parameters:
        - { name: tokenID, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: OK }
...
  /blood/reports/{reportID}:
    parameters:
      - { name: reportID, in: path, required: true, schema: { type: integer } }
    get:
      summary: Read a report and its markers
      responses:
        '200': { description: Success }
    patch:
      summary: Correct report metadata and markers
      responses:
        '200': { description: Success }
    delete:
      summary: Delete a report and its file
      responses:
        '200': { description: Success }
  /blood/reports/{reportID}/file:
    parameters:
      - { name: reportID, in: path, required: true, schema: { type: integer } }
    get:
      summary: Download the original report PDF
      responses:
        '200': { description: Success }
...